connects to database and creates tables

+++ src/main/java/com/example/booklistapp/repository/ReadingListRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ReadingListRepository extends JpaRepository<ReadingList, Integer> {
+public interface ReadingListRepository extends JpaRepository<ReadingList, Long> {
 
 }
